backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.4
alembic==1.14.0
redis==5.2.1
pydantic==2.10.4
email-validator==2.2.0
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
slowapi==0.1.9
httpx==0.28.1
